requirements.txt
Django==5.1.4
django-jazzmin==3.0.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
drf-yasg==1.21.7
datetime==5.5
djangorestframework-simplejwt==5.3.1
requests==2.32.3
pyotp==2.9.0
qrcode==8.0
python-magic==0.4.27
pillow==10.4.0
django-filter==24.3
argon2-cffi==23.1.0
django-redis==5.4.0
django_celery_beat==2.7.0
redis==5.2.0
django-celery-results==2.5.1
openpyxl==3.1.5
celery[redis]==5.4.0
django-environ==0.11.2
uvicorn==0.32.1
gunicorn==23.0.0
psycopg==3.2.3
aiohttp==3.11.9
geoip2==4.8.1
drf-spectacular==0.28.0
msgpack==1.1.0
boto3==1.35.99
django-storages==1.14.4